less max accel at speed
