+++ selfdrive/controls/lib/drive_helpers.py
@@ -1,3 +1,4 @@
+import math
 from cereal import car
 from common.numpy_fast import clip, interp
 from common.realtime import DT_MDL
@@ -18,6 +19,35 @@ CAR_ROTATION_RADIUS = 0.0
 # this corresponds to 80deg/s and 20deg/s steering angle in a toyota corolla
 MAX_CURVATURE_RATES = [0.03762194918267951, 0.003441203371932992]
 MAX_CURVATURE_RATE_SPEEDS = [0, 35]
+
+
+LON_MPC_STEP = 0.2  # first step is 0.2s
+AWARENESS_DECEL = -0.2     # car smoothly decel at .2m/s^2 when user is distracted
+A_CRUISE_MIN = -1.2
+A_CRUISE_MAX_VALS = [1.2, 1.2, 0.8]
+A_CRUISE_MAX_BP = [0., 15., 40.]
+
+# Lookup table for turns
+_A_TOTAL_MAX_V = [1.7, 3.2]
+_A_TOTAL_MAX_BP = [20., 40.]
+
+
+def get_max_accel(v_ego):
+  return interp(v_ego, A_CRUISE_MAX_VALS, A_CRUISE_MAX_VALS)
+
+
+def limit_accel_in_turns(v_ego, angle_steers, a_target, CP):
+  """
+  This function returns a limited long acceleration allowed, depending on the existing lateral acceleration
+  this should avoid accelerating when losing the target in turns
+  """
+
+  a_total_max = interp(v_ego, _A_TOTAL_MAX_BP, _A_TOTAL_MAX_V)
+  a_y = v_ego**2 * angle_steers * CV.DEG_TO_RAD / (CP.steerRatio * CP.wheelbase)
+  a_x_allowed = math.sqrt(max(a_total_max**2 - a_y**2, 0.))
+
+  return [a_target[0], min(a_target[1], a_x_allowed)]
+
 
 class MPC_COST_LAT:
   PATH = 1.0
